yay it works! more stuff to do

+++ src/main/java/daos/BachecaDAO.java
@@ -5,10 +5,7 @@ import databaseConnection.ConnessioneDatabase;
 import model.Bacheca;
 import model.User;
 
-import java.sql.Connection;
-import java.sql.PreparedStatement;
-import java.sql.ResultSet;
-import java.sql.SQLException;
+import java.sql.*;
 import java.util.ArrayList;
 
 public class BachecaDAO {
@@ -21,19 +18,18 @@ public class BachecaDAO {
         }
     }
 
-    public ArrayList<Bacheca> getAllBacheca(User user) {
+    public ArrayList<Bacheca> getAllBacheca(String username) {
         try{
             ArrayList<Bacheca> bacheche = new ArrayList<Bacheca>();
             PreparedStatement recuperaBacheche = connection.prepareStatement("Select * From Bacheca where \"Owner\" = ?");
-            recuperaBacheche.setString(1, user.getUsername());
+            recuperaBacheche.setString(1, username);
             ResultSet resultSet = recuperaBacheche.executeQuery();
             while (resultSet.next()) {
-                bacheche.add(new Bacheca(resultSet.getInt("\"ID\""),
+                bacheche.add(new Bacheca(resultSet.getInt("ID"),
                         resultSet.getString("Title"),
                         resultSet.getString("Description"),
                         resultSet.getBoolean("IsDefault")));
             }
-            connection.close();
             return bacheche;
 
         }
@@ -45,28 +41,51 @@ public class BachecaDAO {
 
     public void createBacheca(Bacheca bacheca, String username) {
         try{
-            PreparedStatement inserisciBacheca = connection.prepareStatement("Insert into Bacheca(Title, Description, \"Owner\") values(?,?,?)");
+            PreparedStatement inserisciBacheca = connection.prepareStatement("Insert into Bacheca(Title, Description, \"Owner\") values(?,?,?)", Statement.RETURN_GENERATED_KEYS);
             inserisciBacheca.setString(1, bacheca.getTitle());
             inserisciBacheca.setString(2, bacheca.getDescription());
             inserisciBacheca.setString(3, username);
+            inserisciBacheca.executeUpdate();
+            ResultSet rs = inserisciBacheca.getGeneratedKeys();
+            rs.next();
+            bacheca.setId(rs.getInt("ID"));
 
-            connection.close();
         }
         catch(SQLException e){
             e.printStackTrace();
         }
     }
 
-    public void deleteBacheca(Bacheca bacheca) {
+    public boolean deleteBacheca(Bacheca bacheca) {
         try{
+            PreparedStatement checkdefault = connection.prepareStatement("Select * from Bacheca where \"ID\" =?");
+            checkdefault.setInt(1, bacheca.getId());
+            ResultSet resultSet = checkdefault.executeQuery();
+            resultSet.next();
+            if (resultSet.getBoolean("isdefault")) {
+                return false;
+            }
             PreparedStatement rimuoviBacheca = connection.prepareStatement("Delete from Bacheca where \"ID\" = ?");
             rimuoviBacheca.setInt(1, bacheca.getId());
-
-            connection.close();
+            rimuoviBacheca.executeUpdate();
+            return true;
         }
         catch(SQLException e){
             e.printStackTrace();
+            return false;
         }
+    }
 
+    public void editBacheca(Bacheca bacheca) {
+        try {
+            PreparedStatement editBacheca = connection.prepareStatement("update bacheca set title = ?," +
+                                                                            "description = ?  where \"ID\" = ? ");
+            editBacheca.setString(1, bacheca.getTitle());
+            editBacheca.setString(2, bacheca.getDescription());
+            editBacheca.setInt(3, bacheca.getId());
+            editBacheca.executeUpdate();
+        } catch (SQLException e) {
+            e.printStackTrace();
+        }
     }
 }
